Using events in Blockservice
