+++ src/main/java/com/example/petbuddybackend/controller/websocket/ChatWebSocketController.java
@@ -6,6 +6,7 @@ import com.example.petbuddybackend.dto.chat.notification.ChatNotificationSend;
 import com.example.petbuddybackend.entity.chat.ChatRoom;
 import com.example.petbuddybackend.entity.user.Role;
 import com.example.petbuddybackend.service.chat.ChatService;
+import com.example.petbuddybackend.service.chat.ChatEventService;
 import com.example.petbuddybackend.service.chat.WebSocketChatMessageSender;
 import com.example.petbuddybackend.service.notification.WebsocketNotificationSender;
 import com.example.petbuddybackend.service.session.chat.ChatSessionTracker;
@@ -51,6 +52,7 @@ public class ChatWebSocketController {
     private final WebSocketChatMessageSender wsChatMessageSender;
     private final WebsocketNotificationSender wsNotificationSender;
     private final ChatSessionTracker chatSessionTracker;
+    private final ChatEventService chatSessionService;
 
     @Transactional
     @MessageMapping("/chat/{chatId}")
@@ -68,7 +70,7 @@ public class ChatWebSocketController {
         ChatRoom chatRoom = chatService.getChatRoomById(chatId);
         String recipientUsername = chatService.getMessageReceiverEmail(principalUsername, chatRoom);
 
-        boolean seenByRecipient = wsChatMessageSender.isRecipientInChat(recipientUsername, chatRoom);
+        boolean seenByRecipient = chatSessionService.isRecipientInChat(recipientUsername, chatRoom);
         ChatMessageDTO messageDTO = chatService.createMessage(
                 chatRoom,
                 principalUsername,
@@ -103,7 +105,7 @@ public class ChatWebSocketController {
         Long chatId = HeaderUtils.getLongFromDestination(accessor, CHAT_ID_INDEX_IN_TOPIC_URL);
 
         chatSessionTracker.addSubscription(accessor.getSubscriptionId(), chatId);
-        wsChatMessageSender.onUserJoinChatRoom(subscriberUsername, chatId);
+        chatSessionService.onUserJoinChatRoom(subscriberUsername, chatId);
         wsNotificationSender.sendNotification(
                 subscriberUsername,
                 chatService.getUnseenChatsNumberNotification(subscriberUsername)
@@ -124,7 +126,7 @@ public class ChatWebSocketController {
         String subId = accessor.getSubscriptionId();
 
         Long chatId = chatSessionTracker.removeSubscription(subId);
-        wsChatMessageSender.onUserUnsubscribe(username, chatId);
+        chatSessionService.onUserUnsubscribe(username, chatId);
 
         log.debug(
                 "Event unsubscribe at {}; sessionId: {}; user: {}",
@@ -140,7 +142,7 @@ public class ChatWebSocketController {
         String username = HeaderUtils.getUser(accessor);
 
         Map<String, Long> subscriptions = chatSessionTracker.getSubscriptions();
-        wsChatMessageSender.onUserDisconnect(username, subscriptions);
+        chatSessionService.onUserDisconnect(username, subscriptions);
         chatSessionTracker.clear();
 
         log.debug(
