--- src/main/java/com/example/petbuddybackend/dto/chat/notification/BlockType.java
@@ -1,5 +1,0 @@
-package com.example.petbuddybackend.dto.chat.notification;
-
-public enum BlockType {
-    BLOCKED, UNBLOCKED
-}
