+++ src/main/java/com/example/petbuddybackend/service/block/BlockService.java
@@ -1,13 +1,10 @@
 package com.example.petbuddybackend.service.block;
 
-import com.example.petbuddybackend.dto.chat.notification.BlockType;
 import com.example.petbuddybackend.dto.user.AccountDataDTO;
 import com.example.petbuddybackend.entity.block.Block;
 import com.example.petbuddybackend.entity.block.BlockId;
 import com.example.petbuddybackend.repository.block.BlockRepository;
-import com.example.petbuddybackend.repository.chat.ChatRoomRepository;
-import com.example.petbuddybackend.service.care.state.CareStateMachine;
-import com.example.petbuddybackend.service.chat.WebSocketChatMessageSender;
+import com.example.petbuddybackend.service.block.event.BlockEvent;
 import com.example.petbuddybackend.service.mapper.UserMapper;
 import com.example.petbuddybackend.service.user.UserService;
 import com.example.petbuddybackend.utils.exception.throweable.general.IllegalActionException;
@@ -16,6 +13,7 @@ import com.example.petbuddybackend.utils.exception.throweable.user.AlreadyBlocke
 import com.example.petbuddybackend.utils.exception.throweable.user.BlockedException;
 import com.example.petbuddybackend.utils.paging.PagingUtils;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -32,11 +30,9 @@ public class BlockService {
     private static final String BLOCK = "Block";
 
     private final BlockRepository blockRepository;
-    private final CareStateMachine careStateMachine;
     private final UserService userService;
-    private final WebSocketChatMessageSender webSocketChatMessageSender;
     private final UserMapper userMapper = UserMapper.INSTANCE;
-    private final ChatRoomRepository chatRoomRepository;
+    private final ApplicationEventPublisher eventPublisher;
 
     public Page<AccountDataDTO> getUsersBlockedByUserSortedByBlockedUsername(String username, Pageable pageable) {
         Pageable sortedPageable = PagingUtils.sortedBy(
@@ -58,16 +54,15 @@ public class BlockService {
         assertNotAlreadyBlocked(blockerUsername, blockedUsername);
 
         blockRepository.save(new Block(blockerUsername, blockedUsername));
-        webSocketChatMessageSender.sendBlockMessageToUsers(blockerUsername, blockedUsername, BlockType.BLOCKED);
-        careStateMachine.cancelCaresIfStatePermitsAndSave(blockerUsername, blockedUsername);
+        eventPublisher.publishEvent(new BlockEvent(blockerUsername, blockedUsername, BlockType.BLOCKED));
     }
 
     public void unblockUser(String blockerUsername, String blockedUsername) {
         assertDoesNotBlockSelf(blockerUsername, blockedUsername);
 
         Block block = getBlock(blockerUsername, blockedUsername);
-        webSocketChatMessageSender.sendBlockMessageToUsers(blockerUsername, blockedUsername, BlockType.UNBLOCKED);
         blockRepository.delete(block);
+        eventPublisher.publishEvent(new BlockEvent(blockerUsername, blockedUsername, BlockType.UNBLOCKED));
     }
 
     public boolean isBlocked(String blockerUsername, String blockedUsername) {
