+++ src/main/java/com/example/petbuddybackend/service/block/BlockType.java
@@ -1,0 +1,5 @@
+package com.example.petbuddybackend.service.block;
+
+public enum BlockType {
+    BLOCKED, UNBLOCKED
+}
