+++ src/main/java/com/example/petbuddybackend/service/block/event/BlockEvent.java
@@ -1,0 +1,10 @@
+package com.example.petbuddybackend.service.block.event;
+
+import com.example.petbuddybackend.service.block.BlockType;
+
+public record BlockEvent(
+        String blockerUsername,
+        String blockedUsername,
+        BlockType blockType
+) {
+}
